+++ rollup.config.js
@@ -1,12 +1,10 @@
 import { terser } from 'rollup-plugin-terser';
 
 export default {
-  input: './src/index.js',
-  output: {
-    file: './dist/strve.esm.js',
-    format: 'esm'
-  },
-  plugins: [
-    terser()
-  ]
-}
+	input: './lib/index.js',
+	output: {
+		file: './dist/strve.esm.js',
+		format: 'esm',
+	},
+	plugins: [terser()],
+};