+++ src/index.js
@@ -1,9 +1,0 @@
-/*!
- * Strve.js v3.0.0
- * (c) 2021-2022 maomincoding
- * Released under the MIT License.
- */
-export { createApp, version } from './init';
-export { watchDom, emit, clone } from './util';
-export { h } from './compile';
-export { setData } from './diff';