+++ src/index.ts
@@ -3,7 +3,11 @@
  * (c) 2021-2022 maomincoding
  * Released under the MIT License.
  */
+
 export { createApp, version } from './init.js';
+
 export { watchDom, emit, clone } from './util.js';
+
 export { h } from './compile.js';
+
 export { setData } from './diff.js';